Chore: Fix dependencies
Hey thanks for the great work, this is really cool, and makes alot of sense. I prefer this over `vuex-typex` as this is more vanilla and more configurable. 

I have moved the packages, to be in devDependencies. This will allow the packages to work long term without need to update.

Also bumped the dependencies to latest vue/vuex versions